// src/app/models/esyaModels/esyaSefer.ts
export interface EsyaSefer {
  Id: number;
  SeferId: number;
  Plaka1: string;
  Plaka2: string;
  Sofor1TCNo: string;
  Sofor2TCNo: string;
  FirmaSeferNo: string;
  BaslangicTarihi: string;
  BaslangicSaati: string;
  BitisTarihi: string;
  BitisSaati: string;
  AktifIptal: string;
  IptalKodu: string;
  IptalAciklama: string;
}